Skipped nodePort service creation for pod network type
