+++ controllers/configmap.go
@@ -8,6 +8,7 @@ import (
 	"fmt"
 	"io/fs"
 	"path/filepath"
+	"strconv"
 	"strings"
 	"text/template"
 
@@ -85,6 +86,15 @@ func init() {
 	if err != nil {
 		// Error reading embedded script templates.
 		panic(fmt.Sprintf("error reading embedded script templates: %v", err))
+	}
+}
+
+func getNamespacedNameForSTSConfigMap(
+	aeroCluster *asdbv1beta1.AerospikeCluster, rackID int,
+) types.NamespacedName {
+	return types.NamespacedName{
+		Name:      aeroCluster.Name + "-" + strconv.Itoa(rackID),
+		Namespace: aeroCluster.Namespace,
 	}
 }
