+++ controllers/pod.go
@@ -11,6 +11,7 @@ import (
 
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/sets"
@@ -251,6 +252,16 @@ func (r *SingleClusterReconciler) restartPods(
 
 	for idx := range podsToRestart {
 		pod := podsToRestart[idx]
+
+		if !podServiceNeeded(r.aeroCluster.Status.PodSpec.MultiPodPerHost, &r.aeroCluster.Status.AerospikeNetworkPolicy) &&
+			podServiceNeeded(r.aeroCluster.Spec.PodSpec.MultiPodPerHost, &r.aeroCluster.Spec.AerospikeNetworkPolicy) {
+			if err := r.createPodService(
+				pod.Name, r.aeroCluster.Namespace,
+			); err != nil && errors.IsAlreadyExists(err) {
+				return reconcileError(err)
+			}
+		}
+
 		// Check if this pod needs restart
 		restartType := restartTypeMap[pod.Name]
 
@@ -549,7 +560,7 @@ func (r *SingleClusterReconciler) cleanupPods(
 			}
 		}
 
-		if r.aeroCluster.Spec.PodSpec.MultiPodPerHost {
+		if podServiceNeeded(r.aeroCluster.Status.PodSpec.MultiPodPerHost, &r.aeroCluster.Status.AerospikeNetworkPolicy) {
 			// Remove service for pod
 			// TODO: make it more robust, what if it fails
 			if err := r.deletePodService(
